+++ Cargo.toml
@@ -24,15 +24,15 @@ members = [
     "datafusion/core",
     "datafusion/expr",
     "datafusion/expr-common",
-    "datafusion/expr-functions-aggregate",
     "datafusion/execution",
+    "datafusion/functions",
     "datafusion/functions-aggregate",
     "datafusion/functions-aggregate-common",
-    "datafusion/functions",
     "datafusion/functions-nested",
     "datafusion/optimizer",
-    "datafusion/physical-expr-common",
     "datafusion/physical-expr",
+    "datafusion/physical-expr-common",
+    "datafusion/physical-expr-functions-aggregate",
     "datafusion/physical-optimizer",
     "datafusion/physical-plan",
     "datafusion/proto",
@@ -98,7 +98,6 @@ datafusion-common-runtime = { path = "datafusion/common-runtime", version = "40.
 datafusion-execution = { path = "datafusion/execution", version = "40.0.0" }
 datafusion-expr = { path = "datafusion/expr", version = "40.0.0" }
 datafusion-expr-common = { path = "datafusion/expr-common", version = "40.0.0" }
-datafusion-expr-functions-aggregate = { path = "datafusion/expr-functions-aggregate", version = "40.0.0" }
 datafusion-functions = { path = "datafusion/functions", version = "40.0.0" }
 datafusion-functions-aggregate = { path = "datafusion/functions-aggregate", version = "40.0.0" }
 datafusion-functions-aggregate-common = { path = "datafusion/functions-aggregate-common", version = "40.0.0" }
@@ -106,6 +105,7 @@ datafusion-functions-nested = { path = "datafusion/functions-nested", version = 
 datafusion-optimizer = { path = "datafusion/optimizer", version = "40.0.0", default-features = false }
 datafusion-physical-expr = { path = "datafusion/physical-expr", version = "40.0.0", default-features = false }
 datafusion-physical-expr-common = { path = "datafusion/physical-expr-common", version = "40.0.0", default-features = false }
+datafusion-physical-expr-functions-aggregate = { path = "datafusion/physical-expr-functions-aggregate", version = "40.0.0" }
 datafusion-physical-optimizer = { path = "datafusion/physical-optimizer", version = "40.0.0" }
 datafusion-physical-plan = { path = "datafusion/physical-plan", version = "40.0.0" }
 datafusion-proto = { path = "datafusion/proto", version = "40.0.0" }
