+++ datafusion/core/src/lib.rs
@@ -556,9 +556,9 @@ pub mod physical_expr_common {
     pub use datafusion_physical_expr_common::*;
 }
 
-/// re-export of [`datafusion_expr_functions_aggregate`] crate
-pub mod expr_functions_aggregate {
-    pub use datafusion_expr_functions_aggregate::*;
+/// re-export of [`datafusion_physical_expr_functions_aggregate`] crate
+pub mod physical_expr_functions_aggregate {
+    pub use datafusion_physical_expr_functions_aggregate::*;
 }
 
 /// re-export of [`datafusion_physical_expr`] crate
